Issue #29: Implementacion de los metodos en Shop.java. Cambio en product para no multiplicar el precio por 2. Cambio en el Dao de Hibernate (y los demas Daos) para eliminar el producto por ID y no por nombre. Cambio en el ProductView para implementar la nueva logica de eliminar por ID del Dao y del Shop.

+++ src/dao/DaoImplHibernate.java
@@ -109,14 +109,20 @@ public class DaoImplHibernate implements Dao {
     }
 
     @Override
-    public void deleteProduct(Product product) {
+    public void deleteProduct(int id) {
         if (session == null || !session.isOpen()) connect();
         Transaction tx = null;
         try {
-            tx = session.beginTransaction();
-            session.delete(product);
+            tx = session.beginTransaction();           
+            Product product = session.get(Product.class, id);
+            
+            if (product != null) {
+                session.delete(product);
+                System.out.println("Producto eliminado con ID: " + id);
+            } else {
+                System.out.println("No se encontró producto con ID: " + id);
+            }            
             tx.commit();
-            System.out.println("Producto eliminado: " + product.getName());
         } catch (Exception e) {
             if (tx != null) tx.rollback();
             e.printStackTrace();
